+++ src/main/java/com/rs/halo/plugin/meilisearch/utils/MeilisearchClientHolder.java
@@ -3,13 +3,16 @@ package com.rs.halo.plugin.meilisearch.utils;
 import com.meilisearch.sdk.Client;
 import com.meilisearch.sdk.Config;
 import com.meilisearch.sdk.Index;
+import com.rs.halo.plugin.meilisearch.config.MeilisearchSetting;
 import com.rs.halo.plugin.meilisearch.event.ConfigUpdatedEvent;
 import lombok.Getter;
+import lombok.extern.slf4j.Slf4j;
 import org.springframework.context.event.EventListener;
 import org.springframework.core.Ordered;
 import org.springframework.core.annotation.Order;
 import org.springframework.stereotype.Component;
 
+@Slf4j
 @Getter
 @Component
 public class MeilisearchClientHolder {
@@ -18,10 +21,11 @@ public class MeilisearchClientHolder {
     private Index index;
 
     @Order(Ordered.HIGHEST_PRECEDENCE)
-    @EventListener
-    public void resetIndex(ConfigUpdatedEvent event) {
+    @EventListener(ConfigUpdatedEvent.class)
+    private void resetIndex() {
+        log.info("reset meilisearch client and index");
         client =
-            new Client(new Config(event.getConfig().getHost(), event.getConfig().getMasterKey()));
+            new Client(new Config(MeilisearchSetting.getHost(), MeilisearchSetting.getMasterKey()));
         index = client.index("halo_post");
     }
 }